feat(logs): introduce LogsCommand trait for enhanced log handling
Refactor the log retrieval mechanism by introducing a new LogsCommand trait, which consolidates the logic for fetching and following Docker container logs. This change improves code modularity and reusability. The execute method now includes a $wait parameter to control process execution, enhancing flexibility. Tests for the new LogsCommand trait ensure reliability and correctness of the log handling functionality.
