+++ src/Docker/Command/LogsCommand.php
@@ -1,0 +1,40 @@
+<?php
+
+namespace Testcontainers\Docker\Command;
+
+use Testcontainers\Docker\Output\DockerFollowLogsOutput;
+use Testcontainers\Docker\Output\DockerLogsOutput;
+use Testcontainers\Docker\Types\ContainerId;
+
+trait LogsCommand
+{
+    /**
+     * Retrieve the logs of a Docker container.
+     *
+     * This method wraps the `docker logs` command to fetch the logs of the specified container.
+     *
+     * @param ContainerId $containerId The ID of the container to fetch logs from.
+     * @param array{
+     *     details?: bool,
+     *     follow?: bool,
+     *     since?: string,
+     *     tail?: string,
+     *     timestamps?: bool,
+     *     until?: string,
+     * } $options Additional options for the Docker logs command.
+     * @return DockerFollowLogsOutput|DockerLogsOutput The output containing the logs of the container.
+     */
+    public function logs($containerId, $options = [])
+    {
+        $follow = isset($options['follow']) ? $options['follow'] : false;
+        $process = $this->execute('logs', null, [(string) $containerId], $options, $follow === false);
+
+        if ($follow === true) {
+            return new DockerFollowLogsOutput($process);
+        } else {
+            return new DockerLogsOutput($process);
+        }
+    }
+
+    abstract protected function execute($command, $subcommand = null, $args = [], $options = [], $wait = true);
+}
