+++ src/Docker/Command/RunCommand.php
@@ -57,5 +57,5 @@ trait RunCommand
         }
     }
 
-    abstract protected function execute($command, $subcommand = null, $args = [], $options = []);
+    abstract protected function execute($command, $subcommand = null, $args = [], $options = [], $wait = true);
 }
