+++ src/Docker/Command/StopCommand.php
@@ -44,5 +44,5 @@ trait StopCommand
         return new DockerStopOutput($process);
     }
 
-    abstract protected function execute($command, $subcommand = null, $args = [], $options = []);
+    abstract protected function execute($command, $subcommand = null, $args = [], $options = [], $wait = true);
 }
